pyproject.toml
[tool.poetry]
name = "plexo"
version = "1.0.0a8"
readme = "README.md"
description = "Opinionated, reactive, schema-driven, distributed, and strongly-typed message passing"
authors = ["Alecks Gates <agates@mail.agates.io>"]
license = "LGPL-3.0-or-later"
homepage = "https://gitlab.com/agates/pyplexo/"
repository = "https://gitlab.com/agates/pyplexo/"
classifiers = [
    "Environment :: Console",
    "Framework :: AsyncIO",
    "Framework :: Pytest",
    "Development Status :: 3 - Alpha",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Adaptive Technologies",
    "Topic :: Home Automation",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
include = [
    { path = "src/plexo/schema/**/*.capnp", format = "sdist"},
    { path = "src/plexo/schema/**/*.py", format = [ "sdist", "wheel" ] },
    { path = "src/plexo/schema/**/*.so", format = "wheel"},
]

[tool.poetry.build]
script = "build.py"
generate-setup-file = true

[tool.poetry.dependencies]
python = "^3.10.0,<3.12"
capnpy = { version = "^0.10.0" }
pyrsistent = "^0.20.0"
pyzmq = "^25.0.0"
returns = "^0.22.0"
typing_extensions = "^4.0"
python-jsonschema-objects = "^0.5.0"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.0"
black = "^23.0.0"
isort = "^5.8.0"
mypy = "^1.0"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.0"
pytest-cov = "^4.0.0"
pytest-mock = "*"
pytest-mypy = "^0.10.0"
pytest-timeout = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0", "capnpy", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py310', 'py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = [
    "tests",
]